+++ backend/backend.config.js
@@ -42,8 +42,8 @@ module.exports = {
         parseInt(process.env.RANKING_POLLING_TIME_MS, 10)
         || 5 * 60 * 1000,
       historySize: 84,
-      erasPerDay: 4,
-      tokenDecimals: 12,
+      erasPerDay: 1,
+      tokenDecimals: 10,
       featuredTimespan: 60 * 60 * 24 * 7 * 2 * 1000, // 2 weeks
     },
     {