+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## vNext
 - ...
 
+## v0.3.0
+- Added faucet functionality.
+- Added Google Analytics for PROD.
+- Updated backend config values.
+
 ## v0.2.0
 - Added UI to show EDP banner on homepage.